Explicit cast buflen to (uint32_t)
netproto_packet.c: we already checked that
    buflen <= UINT32_MAX
(29 lines above), so the cast is safe.

netpacket_write.c: we already checked that
    buflen <= 262144
(which is 2^18) so this cast is safe.

storage_read.c: we already checked that
    buflen <= 262144 - STORAGE_FILE_OVERHEAD
so it's safe to cast (buflen + STORAGE_FILE_OVERHEAD) to (uint32_t).

+++ lib/netpacket/netpacket_write.c
@@ -80,7 +80,7 @@ netpacket_write_file(NETPACKET_CONNECTION * NPC,
 	packetbuf[8] = class;
 	memcpy(&packetbuf[9], name, 32);
 	memcpy(&packetbuf[41], nonce, 32);
-	be32enc(&packetbuf[73], buflen);
+	be32enc(&packetbuf[73], (uint32_t)buflen);
 	memcpy(&packetbuf[77], buf, buflen);
 
 	/* Append hmac. */
